requirements.txt
tensorflow-cpu==2.18.0
numpy==1.26.4
pandas==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
flask==3.0.0
Pillow==10.1.0
joblib==1.3.2
pytest==7.4.0
